requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
h2==4.1.0
PyJWT[crypto]==2.9.0
